% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_label.R
\name{append_label}
\alias{append_label}
\title{Append to \verb{@label} of \linkS4class{panel}}
\usage{
append_label(x, info)
}
\arguments{
\item{x}{\linkS4class{panel}}

\item{info}{\link[base]{character} scalar}
}
\value{
The function \code{\link[=append_label]{append_label()}} returns an \linkS4class{panel} object.
}
\description{
Append to \verb{@label} of \linkS4class{panel}
}
\note{
There is no generic function \verb{labels<-} in package \pkg{base}.

The function \link[base]{comment<-} is not an S3 generic.
}
