% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_panel.R
\name{as_flextable.panel}
\alias{as_flextable.panel}
\title{as_flextable.panel}
\usage{
\method{as_flextable}{panel}(x, ...)
}
\arguments{
\item{x}{a \linkS4class{panel}}

\item{...}{..}
}
\description{
as_flextable.panel
}
\keyword{internal}
