% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panellist.R
\name{as_flextable.panellist}
\alias{as_flextable.panellist}
\title{Convert a \link{panellist} into \link[flextable]{flextable}}
\usage{
\method{as_flextable}{panellist}(x, ...)
}
\arguments{
\item{x}{a \link{panellist}}

\item{...}{..}
}
\description{
Convert a \link{panellist} into \link[flextable]{flextable}
}
\keyword{internal}
