% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panellist.R
\name{autoplot.panellist}
\alias{autoplot.panellist}
\alias{autolayer.panellist}
\title{Visualize \link{panellist}}
\usage{
\method{autoplot}{panellist}(object, ...)

\method{autolayer}{panellist}(object, which = c("oc", "roc"), ...)
}
\arguments{
\item{object}{a \link{panellist}}

\item{...}{additional parameters, currently no use}

\item{which}{\link[base]{character} scalar,
\code{'oc'} (default) or (faux) \code{'roc'}}
}
\description{
Visualize \link{panellist}
}
\keyword{internal}
