% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consort.R
\name{consort.panel}
\alias{consort.panel}
\title{CONSORT Diagram of Ordered \linkS4class{panel}}
\usage{
consort.panel(x, ...)
}
\arguments{
\item{x}{an ordered \linkS4class{panel}}

\item{...}{additional parameters of the function \link[consort]{consort_plot},
except for \code{data}, \code{orders} and \code{side_box}}
}
\value{
The function \code{\link[=consort.panel]{consort.panel()}} returns an R object of class \code{'consort'}.
}
\description{
CONSORT diagram of an ordered \linkS4class{panel}.
}
\keyword{internal}
