% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consort.R
\name{consort.panellist}
\alias{consort.panellist}
\title{CONSORT Diagram of Ordered \link{panellist}}
\usage{
consort.panellist(x, ...)
}
\arguments{
\item{x}{an ordered \link{panellist}}

\item{...}{additional parameters for the function \link[patchwork]{wrap_plots},
\strong{not} for the function \code{\link[=consort.panel]{consort.panel()}}}
}
\value{
The function \code{\link[=consort.panellist]{consort.panellist()}} returns a \link[patchwork]{patchwork}.
}
\description{
CONSORT diagram of an ordered \link{panellist}.
}
