% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{livercancer24}
\alias{livercancer24}
\alias{m1}
\alias{m0}
\title{2024 Liver Cancer paper}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 12877 rows and 50 columns.

An object of class \code{matrix} (inherits from \code{array}) with 12877 rows and 31 columns.
}
\usage{
m1

m0
}
\description{
2024 Liver Cancer paper
}
\keyword{datasets}
\keyword{internal}
