% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{minPanel-package}
\alias{minPanel}
\alias{minPanel-package}
\title{minPanel: Minimum Panel by Sensitivity and Specificity}
\description{
Minimum panel by sensitivity and specificity.
}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com} (\href{https://orcid.org/0000-0001-9971-4844}{ORCID})

}
\keyword{internal}
