% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{ordPanel-package}
\alias{ordPanel}
\alias{ordPanel-package}
\title{ordPanel: Ordered Panel}
\description{
Ordered panel.
}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com} (\href{https://orcid.org/0000-0001-9971-4844}{ORCID})

}
\keyword{internal}
