% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panellist.R
\name{panellist}
\alias{panellist}
\title{\code{panellist}}
\usage{
panellist(...)
}
\arguments{
\item{...}{one or more \linkS4class{panel}s}
}
\value{
The function \code{\link[=panellist]{panellist()}} returns an \code{S3} object of \code{panellist},
which inherits from the classes \link[stats]{listof} and \link[base]{list}.
}
\description{
To combine multiple \linkS4class{panel}s.
}
\keyword{internal}
