% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consort.R
\name{plot.panel}
\alias{plot.panel}
\title{CONSORT-Like Flow-Chart of Ordered \linkS4class{panel}}
\usage{
\method{plot}{panel}(x, ...)
}
\arguments{
\item{x}{an ordered \linkS4class{panel}}

\item{...}{additional parameters of the function \link[consort]{consort_plot},
except for \code{data}, \code{orders} and \code{side_box}}
}
\value{
The function \code{\link[=plot.panel]{plot.panel()}} returns an R object of class \code{'consort'}.
}
\description{
To create a CONSORT-Like flow-chart of an ordered \linkS4class{panel}.
}
