% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consort.R
\name{plot.panellist}
\alias{plot.panellist}
\title{Flow-Charts of Ordered \link{panellist}}
\usage{
\method{plot}{panellist}(x, ...)
}
\arguments{
\item{x}{an ordered \link{panellist}}

\item{...}{additional parameters for the function \link[patchwork]{wrap_plots},
\strong{not} for the function \code{\link[=plot.panel]{plot.panel()}}}
}
\value{
The function \code{\link[=plot.panellist]{plot.panellist()}} returns a \link[patchwork]{patchwork}.
}
\description{
To create flow-charts for the creation of an ordered \link{panellist}.
}
