% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_panel.R
\name{sort_by.panel}
\alias{sort_by.panel}
\title{Sort \linkS4class{panel} by Given Criterion}
\usage{
\method{sort_by}{panel}(x, y, ...)
}
\arguments{
\item{x}{\linkS4class{panel}}

\item{y}{one-sided \link[stats]{formula}}

\item{...}{additional parameters of \link[base]{order}}
}
\value{
The \code{S3} method \code{\link[=sort_by.panel]{sort_by.panel()}} returns a \linkS4class{panel}.
}
\description{
Sort \linkS4class{panel} by Given Criterion
}
