% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_panel.R
\name{[.panel}
\alias{[.panel}
\title{Extract Rows of \linkS4class{panel}}
\usage{
\method{[}{panel}(x, i, ...)
}
\arguments{
\item{x}{\linkS4class{panel}}

\item{i}{\link[base]{logical} \link[base]{vector}, row indices}

\item{...}{additional parameters, currently not in use}
}
\value{
The \code{S3} method \verb{[.panel()} returns a \linkS4class{panel}.
}
\description{
Extract Rows of \linkS4class{panel}
}
\keyword{internal}
