% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_positive.R
\name{sum1}
\alias{sum1}
\alias{sum0}
\alias{cumsum1}
\alias{cumsum0}
\title{Number of True & False Positives, per variant}
\usage{
sum1(x)

sum0(x)

cumsum1(x)

cumsum0(x)
}
\arguments{
\item{x}{\linkS4class{panel}}
}
\value{
The functions \code{\link[=sum1]{sum1()}} and \code{\link[=sum0]{sum0()}} return
a \link[base]{integer} \link[base]{vector}.

The functions \code{\link[=cumsum1]{cumsum1()}} and \code{\link[=cumsum0]{cumsum0()}} return
a \strong{strictly} increasing \link[base]{integer} \link[base]{vector}.
}
\description{
Number of True & False Positives, per variant
}
\keyword{internal}
