% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_panel.R
\name{as_flextable.panel}
\alias{as_flextable.panel}
\title{Convert \linkS4class{panel} into \link[flextable]{flextable}}
\usage{
\method{as_flextable}{panel}(x, ...)
}
\arguments{
\item{x}{\linkS4class{panel}}

\item{...}{additional parameters, currently of no use}
}
\value{
The \code{S3} method \code{\link[=as_flextable.panel]{as_flextable.panel()}} returns a \link[flextable]{flextable}.
}
\description{
To convert an ordered \linkS4class{panel} into a \link[flextable]{flextable}
using the package \CRANpkg{flextable}.
}
\references{
\url{https://tingtingzhan-ordpanel.netlify.app/panel.html}
}
