% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panellist.R
\name{as_flextable.panellist}
\alias{as_flextable.panellist}
\title{Convert a \link{panellist} into \link[flextable]{flextable}}
\usage{
\method{as_flextable}{panellist}(x, ...)
}
\arguments{
\item{x}{\link{panellist}}

\item{...}{additional parameters, currently of no use}
}
\value{
The \code{S3} method \code{\link[=as_flextable.panellist]{as_flextable.panellist()}} returns a \link[flextable]{flextable}.
}
\description{
Convert a \link{panellist} into \link[flextable]{flextable}
}
\note{
The \code{S3} method \code{\link[=as_flextable.panellist]{as_flextable.panellist()}} may be eventually deprecated.
}
\keyword{internal}
