% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{ordPanel-package}
\alias{ordPanel}
\alias{ordPanel-package}
\title{ordPanel: Ordered Panel}
\description{
Ordered panel of variant-signatures by sensitivity and specificity.
}
\note{
Help files of individual functions are intentionally suppressed in the \code{pdf} manual.
Users are encouraged to get started with the vignettes (in the URL's at the bottom of this documentation).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tingtingzhan/ordPanel}
  \item \url{https://tingtingzhan-ordpanel.netlify.app}
}

}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com} (\href{https://orcid.org/0000-0001-9971-4844}{ORCID})

Other contributors:
\itemize{
  \item Aejaz Sayeed \email{aejaz.sayeed@bblumberg.org} (\href{https://orcid.org/0000-0001-6900-7524}{ORCID}) [contributor]
}

}
