% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{ordPanel-package}
\alias{ordPanel}
\alias{ordPanel-package}
\title{ordPanel: Ordered Panel}
\description{
The ordered panel methodology (Zezulinski et al 2025 \doi{10.1159/000545366}) provides a structured framework for identifying and organizing sets of biomarkers, such as genetic variants, that distinguish between positive and negative subjects in a study when only a training cohort is available. This approach is particularly useful in situations where an independent validation cohort does not yet exist, rendering conventional performance metrics such as the receiver operating characteristic (ROC) curve and area under the ROC curve (AUC) inappropriate or potentially misleading. The methodology emphasizes transparent construction and evaluation of ordered signatures of biomarkers, allowing investigators to examine operating characteristics without establishing predictive performance.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tingtingzhan/ordPanel}
  \item \url{https://tingtingzhan-ordpanel.netlify.app}
}

}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com} (\href{https://orcid.org/0000-0001-9971-4844}{ORCID})

Other contributors:
\itemize{
  \item Aejaz Sayeed (\href{https://orcid.org/0000-0001-6900-7524}{ORCID}) [contributor]
}

}
