% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibs.R
\name{ordPanel_bib}
\alias{ordPanel_bib}
\alias{.zezulinski25}
\title{Bibliography in package \pkg{ordPanel}}
\usage{
.zezulinski25(key = "Zezulinski25", ...)
}
\arguments{
\item{key, ...}{additional parameters of the function \link[utils]{bibentry}}
}
\value{
The function \code{\link[=.zezulinski25]{.zezulinski25()}} returns a \link[utils]{bibentry} object.
}
\description{
Bibliography in package \pkg{ordPanel}
}
