% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.R
\docType{class}
\name{panel}
\alias{panel}
\alias{panel-class}
\title{\code{S4} Class \linkS4class{panel}}
\usage{
panel(m1, m0)
}
\arguments{
\item{m1, m0}{see \strong{Slots}}
}
\value{
The function \code{\link[=panel]{panel()}} returns an R object of \code{S4} class \linkS4class{panel}.
}
\description{
\code{S4} Class \linkS4class{panel}
}
\section{Slots}{

\describe{
\item{\code{m1}}{\link[base]{logical} \link[base]{matrix}, true positives,
i.e., variants tested positive in the positive subjects (patients).
Rows are different variants.  Columns are different subjects.
The \link[base]{rownames} of \code{m0} and \code{m1} must be the same}

\item{\code{m0}}{\link[base]{logical} \link[base]{matrix}, false positives,
i.e., variants tested positive in the negative subjects (patients).
Rows are different variants.  Columns are different subjects.
The \link[base]{rownames} of \code{m0} and \code{m1} must be the same.}

\item{\code{id}}{\link[base]{list} of \link[base]{character} \link[base]{vector}s}

\item{\code{label}}{(optional) \link[base]{character} scalar,
a human-friendly description of the \linkS4class{panel}}

\item{\code{consort}}{(optional) \link[base]{data.frame} to
create a \link[consort]{consort_plot}}
}}

\keyword{internal}
