% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.R
\docType{class}
\name{panel}
\alias{panel}
\alias{panel-class}
\title{\code{S4} Class \linkS4class{panel}}
\usage{
panel(m1 = zezulinski1, m0 = zezulinski0)
}
\arguments{
\item{m1, m0}{see detailed explanations in Section \strong{Slots}.
\itemize{
\item{If both \code{m1} and \code{m0} are missing, then random \link[base]{logical} \link[base]{matrix}-es
will be generated;}
\item{If one-and-only-one of \code{m1} and \code{m0} is missing, then the function \code{\link[=panel]{panel()}} will \link[base]{stop}.}
}}
}
\value{
The function \code{\link[=panel]{panel()}} returns an R object of \code{S4} class \linkS4class{panel}.
}
\description{
\code{S4} Class \linkS4class{panel}
}
\section{Slots}{

\describe{
\item{\code{m1,m0}}{\link[base]{logical} \link[base]{matrix}-es, true and false positives, respectively.
In other words, the variants tested positive in the positive and negative subjects (patients), respectively.
Rows are different variants.  Columns are different subjects.
The \link[base]{rownames} of \code{m0} and \code{m1} must be the same.}

\item{\code{id}}{\link[base]{list} of \link[base]{character} \link[base]{vector}s}

\item{\code{ordered}}{\link[base]{logical} scalar, whether this is an ordered \linkS4class{panel}}

\item{\code{label}}{(optional) \link[base]{character} scalar,
a human-friendly description of the \linkS4class{panel}}

\item{\code{consort}}{(optional) \link[base]{data.frame} to
create a \link[consort]{consort_plot}}
}}

