% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{pkg_data}
\alias{pkg_data}
\alias{m1}
\alias{m0}
\title{Data Sets in Package \pkg{ordPanel}}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 12877 rows and 50 columns.

An object of class \code{matrix} (inherits from \code{array}) with 12877 rows and 31 columns.
}
\usage{
m1

m0
}
\description{
Data sets used as examples in package \pkg{ordPanel}.
}
\examples{
# data used in 
.zezulinski25()


}
\references{
\doi{10.1159/000545366}
}
\keyword{datasets}
