% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.R
\name{show,panel-method}
\alias{show,panel-method}
\title{Show \linkS4class{panel}}
\usage{
\S4method{show}{panel}(object)
}
\arguments{
\item{object}{\linkS4class{panel}}
}
\description{
Show \linkS4class{panel}
}
