% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_panel.R
\name{subset.panel}
\alias{subset.panel}
\title{Select a \link[base]{subset} of \linkS4class{panel}}
\usage{
\method{subset}{panel}(x, subset, append.label = FALSE, ...)
}
\arguments{
\item{x}{\linkS4class{panel}}

\item{subset}{R \link[base]{language} object}

\item{append.label}{\link[base]{logical} scalar (default value \code{FALSE}),
whether to append the subset-criterion to \code{x@label}}

\item{...}{additional parameters, currently not in use}
}
\value{
The \code{S3} method \code{\link[=subset.panel]{subset.panel()}} returns an R object of \code{S4} class \linkS4class{panel}.
}
\description{
Select a \link[base]{subset} of \linkS4class{panel}
}
\keyword{internal}
